// frontend/projects/modeling/src/lib/api/generated/dms/model/typeDefinitionDTOObject.ts
/**
 * PLASMA DMS Service
 * Service for modeling data sources
 *
 * The version of the OpenAPI document: 1.3.0
 * Contact: pomp@uni-wuppertal.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TypeDefinitionDTOObject { 
    type?: string;
    name?: string;
    label?: string;
    description?: string;
    minCardinality?: number;
    maxCardinality?: number;
    value?: Array<object>;
    hidden?: boolean;
}